Course UI - style refactoring, css module abstraction, integrating sort component
