+++ src/app/components/course/CourseSearch/CourseSearchError.tsx
@@ -3,7 +3,7 @@
 import { BUTTON_STYLE } from '@/app/utilities/constants';
 import ActionButton from '../../buttons/ActionButton';
 import Typography, { TypographyVariant } from '../../typography/Typography';
-import styles from './courseSearchResult.module.css';
+import styles from './courseSearchError.module.css';
 
 export default function CourseSearchError({
   reset,
@@ -11,7 +11,7 @@ export default function CourseSearchError({
   reset: () => void;
 }): React.ReactNode {
   return (
-    <div role='alert' className={styles.alert} aria-atomic>
+    <div role='alert' className={styles.container} aria-atomic>
       <Typography variant={TypographyVariant.H3}>
         Failed to fetch the list of courses.
       </Typography>
