+++ src/app/components/course/CourseSearch/courseSearch.module.css
@@ -14,6 +14,21 @@
   gap: 80px;
 }
 
+.container *[role='combobox'] :global(.dropdown-input-wrapper) {
+  border: 0;
+  margin-left: -10px;
+  width: calc(100% + 10px);
+}
+
+.container *[role='combobox'] :global(.dropdown-input-wrapper) > *:first-child {
+  padding-left: 10px;
+}
+
+.container *[role='combobox'] input[role='searchbox'] {
+  padding-left: 0px;
+  margin-left: 0px;
+}
+
 @media (max-width: 768px) {
   .body {
     gap: 40px;
