+++ src/app/components/course/CourseSearch/courseSearchEmpty.module.css
@@ -1,0 +1,14 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  align-items: center;
+}
+
+.text {
+  font-size: 20px;
+  font-weight: 600;
+  line-height: 30px;
+  text-align: center;
+  color: var(--bond-black-light);
+}
