+++ src/app/components/course/CourseSearch/courseSearchError.module.css
@@ -1,0 +1,6 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  align-items: center;
+}
